feat(pages): mainpageにProfileコンポーネントを追加
