+++ frontend/src/pages/PortfolioMainPage/style.scss
@@ -1,0 +1,3 @@
+.container {
+  margin: 32px;
+}
